// src/integrations/cms/generated/cms-schema.d.ts
export interface components {
    schemas: {
// ...
        ApiPromotionPromotionDocument: {
// ...
            content: unknown[];
// ...
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    "product-category/get/product_categories_by_id": {
        parameters: {
            query?: {
                fields?: ("name" | "slug" | "description" | "sortOrder" | "isActive" | "publishedDate" | "createdAt" | "updatedAt" | "publishedAt")[];
                populate?: "*" | ("icon" | "bannerImage" | "seo" | "ProductCategory" | "promotions") | ("icon" | "bannerImage" | "seo" | "ProductCategory" | "promotions")[];
                filters?: {
                    [key: string]: unknown;
                };
                sort?: ("name" | "slug" | "description" | "sortOrder" | "isActive" | "publishedDate" | "createdAt" | "updatedAt" | "publishedAt") | ("name" | "slug" | "description" | "sortOrder" | "isActive" | "publishedDate" | "createdAt" | "updatedAt" | "publishedAt")[] | {
                    [key: string]: "asc" | "desc";
                } | {
                    [key: string]: "asc" | "desc";
                }[];
            };
            header?: never;
            path: {
                id: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        data: {
                            /**
                             * Format: uuid
                             * @description The document ID, represented by a UUID
                             */
                            documentId: string;
                            id: string | number;
                            /** @description A string field */
                            name: string;
                            /** @description A UID field */
                            slug: string;
                            /** @description A text field */
                            description?: string;
                            /**
                             * @description An integer field
                             * @default 0
                             */
                            sortOrder: number;
                            /**
                             * @description A boolean field
                             * @default true
                             */
                            isActive: boolean | null;
                            /** @description A datetime field */
                            publishedDate?: string;
                            /** @description A datetime field */
                            createdAt?: string;
                            /** @description A datetime field */
                            updatedAt?: string;
                            /**
                             * @description A datetime field
                             * @default 2026-06-04T06:49:03.909Z
                             */
                            publishedAt: string;
                            /** @description A media field */
                            icon?: components["schemas"]["PluginUploadFileDocument"];
                            /** @description A media field */
                            bannerImage?: components["schemas"]["PluginUploadFileDocument"][];
                            /** @description A component field */
                            seo?: components["schemas"]["SharedSeoEntry"];
                            /** @description A relational field */
                            ProductCategory?: components["schemas"]["ApiProductProductDocument"][];
                            /** @description A relational field */
                            promotions?: components["schemas"]["ApiPromotionPromotionDocument"][];
                        };
                    };
                };
            };
            /** @description Bad request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Unauthorized */
            401: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Forbidden */
            403: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Not found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Internal server error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
// ...
}